+++ GTiff2Tiles.Benchmarks/Program.cs
@@ -1,234 +1,34 @@
-﻿using System;
-using System.Diagnostics;
-using System.IO;
-using System.Reflection;
-using System.Threading.Tasks;
+﻿using System.IO;
+using BenchmarkDotNet.Running;
 using CommandLine;
-using GTiff2Tiles.Core.Enums;
-using GTiff2Tiles.Core.GeoTiffs;
-
-// ReSharper disable LocalizableElement
-// ReSharper disable UnusedParameter.Local
 
 namespace GTiff2Tiles.Benchmarks
 {
     internal static class Program
     {
-        #region Properties
+        private static string _inputPath = "../../../../../Examples/Input/Benchmark.tif";
 
-        /// <summary>
-        /// Input file.
-        /// </summary>
-        private static FileInfo InputFileInfo { get; set; }
+        private static bool _isParsingErrors;
 
-        /// <summary>
-        /// Output directory.
-        /// </summary>
-        private static DirectoryInfo OutputDirectoryInfo { get; set; }
-
-        /// <summary>
-        /// Temp directory.
-        /// </summary>
-        private static DirectoryInfo TempDirectoryInfo { get; set; }
-
-        /// <summary>
-        /// Minimum cropped zoom.
-        /// </summary>
-        private static int MinZ { get; set; }
-
-        /// <summary>
-        /// Maximum cropped zoom.
-        /// </summary>
-        private static int MaxZ { get; set; }
-
-        /// <summary>
-        /// Shows if there were console line parsing errors.
-        /// </summary>
-        private static bool IsParsingErrors { get; set; }
-
-        /// <summary>
-        /// Threads count.
-        /// </summary>
-        private static int ThreadsCount { get; set; }
-
-        #endregion
-
-        internal static async Task Main(string[] args)
+        private static void Main(string[] args)
         {
-            Stopwatch stopwatch = new Stopwatch();
-            stopwatch.Start();
-
             Parser.Default.ParseArguments<Options>(args).WithParsed(ParseConsoleOptions)
-                  .WithNotParsed(error => IsParsingErrors = true);
+                  .WithNotParsed(error => _isParsingErrors = true);
 
-            if (IsParsingErrors) throw new Exception("Error while parsing occured");
+            if (_isParsingErrors) return;
 
-            await RunTilingAsync().ConfigureAwait(false);
+            Directory.CreateDirectory($"{Benchmarks.Data}/{Benchmarks.In}");
+            // Copy benchmark.tif to data/in/i.tif
+            File.Copy(_inputPath, $"{Benchmarks.Data}/{Benchmarks.In}/{Benchmarks.Itif}", true);
 
-            Console.WriteLine("Tests ended. Press any button to close the application.");
-            Console.ReadKey();
+            // Pull the latest images
+            Docker.Pull(Benchmarks.GdalName);
+            Docker.Pull(Benchmarks.MaptilerName);
+
+            // Run the benchmarks
+            BenchmarkRunner.Run<Benchmarks>();
         }
 
-        private static void ParseConsoleOptions(Options options)
-        {
-            //Check if string options are empty strings.
-            if (string.IsNullOrWhiteSpace(options.InputFilePath))
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            if (string.IsNullOrWhiteSpace(options.OutputDirectoryPath))
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            //Check zooms.
-            if (options.MinZ < 0)
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            if (options.MaxZ < 0)
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            if (options.MinZ > options.MaxZ)
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            //Threads check.
-            if (options.ThreadsCount <= 0)
-            {
-                IsParsingErrors = true;
-
-                return;
-            }
-
-            //Set properties values.
-            InputFileInfo = new FileInfo(options.InputFilePath);
-            OutputDirectoryInfo = new DirectoryInfo(options.OutputDirectoryPath);
-            TempDirectoryInfo = string.IsNullOrWhiteSpace(options.TempDirectoryPath)
-                                    ? new FileInfo(Assembly.GetExecutingAssembly().Location).Directory
-                                    : new DirectoryInfo(options.TempDirectoryPath);
-            MinZ = options.MinZ;
-            MaxZ = options.MaxZ;
-            ThreadsCount = options.ThreadsCount;
-        }
-
-        private static async ValueTask RunTilingAsync()
-        {
-            CoordinateSystem coordinateSystem = CoordinateSystem.Epsg4326;
-
-            Stopwatch stopwatch = new Stopwatch();
-            stopwatch.Start();
-
-            #region GTiff2Tiles
-
-            string gtiff2TilesOutputDirectoryPath = Path.Combine(OutputDirectoryInfo.FullName, "gtiff2tiles");
-            string gtiff2TilesTempDirectoryPath = Path.Combine(TempDirectoryInfo.FullName, "gtiff2tiles");
-            Directory.CreateDirectory(gtiff2TilesOutputDirectoryPath);
-            Directory.CreateDirectory(gtiff2TilesTempDirectoryPath);
-
-            await using Raster image = new Raster(InputFileInfo?.FullName, coordinateSystem);
-
-            // Generate tiles
-            await image.WriteTilesToDirectoryAsync(gtiff2TilesOutputDirectoryPath, MinZ, MaxZ, true,
-                                                   tileExtension: TileExtension.Png,
-                                                   bandsCount: 4,
-                                                   threadsCount: ThreadsCount)
-                       .ConfigureAwait(false);
-
-            stopwatch.Stop();
-            Console.WriteLine("GTiff2Tiles process ended.");
-            Console.WriteLine("Time passed:");
-            Console.WriteLine($"Hours:{stopwatch.Elapsed.Hours}, Minutes:{stopwatch.Elapsed.Minutes}, " +
-                              $"Seconds:{stopwatch.Elapsed.Seconds}, Ms:{stopwatch.Elapsed.Milliseconds}.");
-            stopwatch.Restart();
-
-            #endregion
-
-            #region MapTiler
-
-            string maptilerOutputDirectoryPath = Path.Combine(OutputDirectoryInfo.FullName, "maptiler");
-            string maptilerTempDirectoryPath = Path.Combine(TempDirectoryInfo.FullName, "maptiler");
-            Directory.CreateDirectory(maptilerOutputDirectoryPath);
-            Directory.CreateDirectory(maptilerTempDirectoryPath);
-            await Task.Run(() => RunMapTiler(InputFileInfo.FullName, maptilerOutputDirectoryPath,
-                                             maptilerTempDirectoryPath, MinZ, MaxZ, ThreadsCount))
-                      .ConfigureAwait(false);
-
-            stopwatch.Stop();
-            Console.WriteLine("MapTiler process ended.");
-            Console.WriteLine("Time passed:");
-            Console.WriteLine($"Hours:{stopwatch.Elapsed.Hours}, Minutes:{stopwatch.Elapsed.Minutes}, " +
-                              $"Seconds:{stopwatch.Elapsed.Seconds}, Ms:{stopwatch.Elapsed.Milliseconds}.");
-            stopwatch.Restart();
-
-            #endregion
-
-            #region Gdal2Tiles
-
-            string gdal2TilesOutputDirectoryPath = Path.Combine(OutputDirectoryInfo.FullName, "gdal2tiles");
-            Directory.CreateDirectory(gdal2TilesOutputDirectoryPath);
-            await Task.Run(() => RunGdal2Tiles(InputFileInfo.FullName, gdal2TilesOutputDirectoryPath, MinZ, MaxZ /*,
-                                               ThreadsCount*/)).ConfigureAwait(false);
-
-            stopwatch.Stop();
-            Console.WriteLine("Gdal2Tiles process ended.");
-            Console.WriteLine("Time passed:");
-            Console.WriteLine($"Hours:{stopwatch.Elapsed.Hours}, Minutes:{stopwatch.Elapsed.Minutes}, " +
-                              $"Seconds:{stopwatch.Elapsed.Seconds}, Ms:{stopwatch.Elapsed.Milliseconds}.");
-
-            #endregion
-        }
-
-        private static void RunMapTiler(string inputFilePath, string outputDirectoryPath, string tempDirectoryPath,
-                                        int minZ, int maxZ, int threadsCount)
-        {
-            using Process maptilerProcess = new Process
-            {
-                StartInfo = new ProcessStartInfo("maptiler")
-                {
-                    Arguments = $"-geodetic -tms -resampling cubic_spline -f png32 -P {threadsCount} " +
-                                $"-o \"{outputDirectoryPath}\" -work_dir \"{tempDirectoryPath}\" " +
-                                $"-srs EPSG:4326 -zoom {minZ} {maxZ} \"{inputFilePath}\"",
-                    CreateNoWindow = true, RedirectStandardInput = true, RedirectStandardOutput = true, UseShellExecute = false
-                }
-            };
-
-            maptilerProcess.Start();
-            maptilerProcess.WaitForExit();
-        }
-
-        private static void RunGdal2Tiles(string inputFilePath, string outputDirectoryPath, int minZ, int maxZ /*,
-            int threadsCount*/)
-        {
-            using Process gdal2TilesProcess = new Process
-            {
-                //Should be placed in Gdal2Tiles directory near Benchmarks binaries.
-                StartInfo = new ProcessStartInfo("Gdal2Tiles/Gdal2Tiles")
-                {
-                    Arguments = $"-s EPSG:4326 -p geodetic -r cubicspline --tmscompatible -z {minZ}-{maxZ} " +
-                                //$"--processes={threadsCount} " + //TODO: doesn't work
-                                $"\"{inputFilePath}\" \"{outputDirectoryPath}\"",
-                    CreateNoWindow = true, RedirectStandardInput = true, RedirectStandardOutput = true, UseShellExecute = false
-                }
-            };
-
-            gdal2TilesProcess.Start();
-            gdal2TilesProcess.WaitForExit();
-        }
+        private static void ParseConsoleOptions(Options options) => _inputPath = string.IsNullOrWhiteSpace(options.InputFilePath) ? _inputPath : options.InputFilePath;
     }
 }