+++ GTiff2Tiles.Benchmarks/Start.ps1
@@ -1,6 +1,9 @@
 # Simple script to run GTiff2Tiles.Benchmarks application, using PowerShell or Powershell.Core on any system.
-# Place this script near GTiff2Tiles.Benchmarks file and uncomment needed lines.
-# Read full docs for application in GTiff2Tiles.Benchmarks.Doc.md or GTiff2Tiles.Benchmarks.Doc.pdf files.
+# Place this script near GTiff2Tiles.Benchmarks.exe or GTiff2Tiles.Benchmarks file and uncomment needed lines.
+# Be sure to read docs for application on github pages: https://gigas002.github.io/GTiff2Tiles/benchmarks-index.html
 
-# Run tiling from Input.tif file to Output directory for 8-11 zooms, crop algorythm, tms-compatible, in 3-threads mode.
-# ./GTiff2Tiles.Benchmarks -i "D:/Examples/Input.tif" -o "D:/Examples/Output" -t "D:/Examples/Temp" --minz 8 --maxz 11 --threads 3
+# Run tiling of custom tif:
+# ./GTiff2Tiles.Benchmarks -i "D:/Examples/Input.tif"
+
+# Run tiling of Examples/Input/Benchmarks.tif
+# ./GTiff2Tiles.Benchmarks